Rename radio C and H files
